EndPoints_Marcelo
AgrupacionesSIndicales(3)
CentrodeCosto
Convenios
Empresas
Especialidades
EstadosUnidades

+++ DATA/Models/Unidades.cs
@@ -57,7 +57,7 @@ namespace DATA.Models
         public virtual Modelos IdModelo { get; set; }
 
         [ForeignKey("idEstadoUnidad")]
-        public int idEstadoUnidad { get; set; }
+        public long idEstadoUnidad { get; set; }
         public virtual EstadosUnidades IdEstadoUnidad { get; set; }
         
         [ForeignKey("idSituacionUnidad")]
